Fix excessive rebuilds of nextjs
Use a build manifest sentinel file for the output of the
nextjs build rule.

Tidy nextjs build/export by wrapping inside a bazel macro
along with the go generation code for the embedded filesystem.

The end result of this is that iterating on go code is much
faster because there are no longer spurious rebuilds of nextjs
when no nextjs source code has changed.

Signed-off-by: Elliot Murphy <statik@users.noreply.github.com>